md5 done, time to sha256

+++ md5basic.c
@@ -15,7 +15,7 @@ uint32_t	h(uint32_t x, uint32_t y, uint32_t z)
 	return (x ^ y ^ z);
 }
 
-uint64_t	i(uint64_t x, uint64_t y, uint64_t z)
+uint32_t	i(uint32_t x, uint32_t y, uint32_t z)
 {
 	return (y ^ (x | ~z));
 }
